Dialer: Whitelist few EU countries for call recording
All entries to this list patch should state the country being modified
and
provide a URL reference (either government law, precedent or similar).

A disclaimer item will also be added to the InCallUI view for the user
to
acknowledge that it is their responsibility to conduct/maintain and use
recordings in conjunction with their respective country/region's laws.

Change-Id: I6677bbf7d97efba3efe9447df83c95ce1162ee33